requirements.dev.txt
-r ./VMAP/requirements.txt

locust
tqdm
ruff
tblib
coverage
django-extensions
django-debug-toolbar
